refactor: tidy up benchmark
