+++ deno.json
@@ -4,8 +4,7 @@
   "license": "MIT",
   "exports": "./index.ts",
   "tasks": {
-    "build": "deno run --allow-env --allow-read --allow-write --allow-run=npm build.ts",
-    "benchmark": "deno bench benchmark.ts"
+    "build": "deno run --allow-env --allow-read --allow-write --allow-run=npm build.ts"
   },
   "publish": {
     "include": [
